Fix task_group sorting by utilizing get_task_group_children_getter to sort based on webserver config
